Fix LaserScan message timestamp
According to the description of `sensor_msgs/LaserScan` [1], "the
timestamp in the header is the acquisition time of the first ray in
the scan".  In both `hlds_laser_publisher.cpp` and
`hlds_laser_segment_publisher.cpp`, however, the timestamp is taken
after the full message has been read, processed, and stored in the
`LaserScan` message, i.e. only after the last ray in the scan.
Consequently, with the lidar spinning at 5rps, the timestamp in the
message is about 0.2s later than it should be.

Acquire the timestamp right after the start sequence has been found to
make the message timestamp more accurate.

[1] https://docs.ros.org/en/noetic/api/sensor_msgs/html/msg/LaserScan.html